+++ examples/1. basic-datatable.blade.php
@@ -1,0 +1,35 @@
+@section('content')
+<x-dg-card title="Release History" bg="primary">
+    <x-dg-datatable id="data_table" :heads="['#', 'Date', 'Hospital', 'Amount']"/>
+</x-dg-card>
+@stop
+
+@section('js')
+<script>
+    $(()=>{
+        data_table = $('#data_table').DataTable({
+            ajax: {
+                url: "{{route('admin.balances.index')}}",
+                type: 'GET',
+                dataSrc: 'data'
+            },
+            columns : [
+                {data: null, defaultContent : ''},
+                {data: 'date'},
+                {data: 'hospital'},
+                {data: 'amount'},
+            ],
+            order: [],
+            autoWidth: false,
+        });
+        data_table.on('draw.dt', function () {
+            var PageInfo = $('#data_table').DataTable().page.info();
+            data_table.column(0, {
+                page: 'all'
+            }).nodes().each(function (cell, i) {
+                cell.innerHTML = i + 1;
+            });
+        });
+    });
+</script>
+@stop