+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 2.1.2
+- updated DataTable component
+- added Print/Export Buttons
+- added some examples in documentation
+
+## 2.1.1
+- updated components
+
+## 2.1.0
+- Added Dynamic Bindings
+
+## 2.0.2
+- Added some html5 sttributes to
+
+## 2.0.1
+- fixed text editor bug
+
+## 2.0.0 
+- Added Plugins and Input Slider
+
 ## 1.0.2
 - Fixed number_format bug